+++ tests/API/ShowTest.php
@@ -1,20 +1,18 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\API;
 
 use KRLX\Show;
-use KRLX\User;
 use KRLX\Term;
 use KRLX\Track;
-use Tests\TestCase;
+use Tests\API\APITestCase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 
-class ShowTest extends TestCase
+class ShowTest extends APITestCase
 {
     use RefreshDatabase;
 
-    public $user;
     public $show;
     public $term;
     public $track;
@@ -23,7 +21,6 @@ class ShowTest extends TestCase
     {
         parent::setUp();
 
-        $this->user = factory(User::class)->create();
         $this->term = factory(Term::class)->create(['accepting_applications' => true]);
         $this->track = factory(Track::class)->create(['active' => true]);
         $this->show = factory(Show::class)->create([
@@ -33,31 +30,15 @@ class ShowTest extends TestCase
     }
 
     /**
-     * Assert that authentication is required to make calls on the API.
-     *
-     * @return void
-     */
-    public function testUnauthenticatedCallsNotPermitted()
-    {
-        $this->assertGuest('api');
-
-        $request = $this->json('POST', '/api/v1/shows', [
-            'title' => 'Gray Duck',
-            'track_id' => $this->track->id,
-            'term_id' => $this->term->id
-        ]);
-
-        $request->assertStatus(401);
-    }
-
-    /**
-     * Assert that shows can be created via the API.
+     * Assert that shows can be created via the API... when signed in.
      *
      * @return void
      */
     public function testShowsCanBeCreated()
     {
-        $request = $this->actingAs($this->user, 'api')->json('POST', '/api/v1/shows', [
+        $this->assertAuthenticatedAs($this->user, 'api');
+
+        $request = $this->json('POST', '/api/v1/shows', [
             'title' => 'Gray Duck',
             'track_id' => $this->track->id,
             'term_id' => $this->term->id
@@ -78,7 +59,7 @@ class ShowTest extends TestCase
      */
     public function testShowsCanGenerateTitles()
     {
-        $request = $this->actingAs($this->user, 'api')->json('POST', '/api/v1/shows', [
+        $request = $this->json('POST', '/api/v1/shows', [
             'track_id' => $this->track->id,
             'term_id' => $this->term->id
         ]);
