+++ tests/API/TermTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\API;
 
 use KRLX\Term;
 use Carbon\Carbon;
